fix REPL e2e test
