+++ e2e/old_repl_test.go
@@ -5,14 +5,12 @@ import (
 	"flag"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
 	"testing"
 
 	"github.com/fatih/color"
-	"github.com/google/go-cmp/cmp"
 	"github.com/ktr0731/evans/app"
 	"github.com/ktr0731/evans/cui"
 	"github.com/ktr0731/evans/prompt"
@@ -376,31 +374,3 @@ var (
 		"\r", "", // For Windows.
 	)
 )
-
-func compareWithGolden(t *testing.T, actual string) {
-	name := t.Name()
-	normalizeFilename := func(name string) string {
-		fname := goldenPathReplacer.Replace(strings.ToLower(name)) + ".golden"
-		return filepath.Join("testdata", "fixtures", fname)
-	}
-
-	fname := normalizeFilename(name)
-
-	if *update {
-		if err := ioutil.WriteFile(fname, []byte(actual), 0600); err != nil {
-			t.Fatalf("failed to update the golden file: %s", err)
-		}
-		return
-	}
-
-	// Load the golden file.
-	b, err := ioutil.ReadFile(fname)
-	if err != nil {
-		t.Fatalf("failed to load a golden file: %s", err)
-	}
-	expected := goldenReplacer.Replace(string(b))
-
-	if diff := cmp.Diff(expected, actual); diff != "" {
-		t.Errorf("wrong result: \n%s", diff)
-	}
-}
