--- e2e/testdata/fixtures/teste2e_cli-call_failure_unary_rpc_with_--print_flag_(all).golden
@@ -1,8 +1,0 @@
-content-type: application/grpc
-header_key1: header_val1
-header_key2: header_val2
-
-trailer_key1: trailer_val1
-trailer_key2: trailer_val2
-
-code = Internal, number = 13, message = "internal error"
