--- e2e/testdata/fixtures/teste2e_cli-call_failure_unary_rpc_with_--print_flag_(messagetrailerstatus)_against_to_grpc-web_server.golden
@@ -1,4 +1,0 @@
-trailer_key1: trailer_val1
-trailer_key2: trailer_val2
-
-code = Internal, number = 13, message = "internal error"
