--- e2e/testdata/fixtures/teste2e_cli-call_unary_rpc_with_--print_flag_(header).golden
@@ -1,3 +1,0 @@
-content-type: application/grpc
-header_key1: header_val1
-header_key2: header_val2
