--- e2e/testdata/fixtures/teste2e_cli-call_unary_rpc_with_--print_flag_(headermessagestatus).golden
@@ -1,9 +1,0 @@
-content-type: application/grpc
-header_key1: header_val1
-header_key2: header_val2
-
-{
-  "message": "response"
-}
-
-code = OK, number = 0, message = ""
