--- e2e/testdata/fixtures/teste2e_cli-call_unary_rpc_with_--print_flag_(message).golden
@@ -1,3 +1,0 @@
-{
-  "message": "response"
-}
