--- e2e/testdata/fixtures/teste2e_cli-call_unary_rpc_with_--print_flag_(status).golden
@@ -1,1 +1,0 @@
-code = OK, number = 0, message = ""
