--- e2e/testdata/fixtures/teste2e_cli-call_unary_rpc_with_--print_flag_(trailer).golden
@@ -1,2 +1,0 @@
-trailer_key1: trailer_val1
-trailer_key2: trailer_val2
